+++ aws-ssm-parameter/src/main/java/com/amazonaws/ssm/parameter/CreateHandler.java
@@ -1,12 +1,13 @@
-package software.amazon.ssm.parameter;
+package com.amazonaws.ssm.parameter;
 
 import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
 import software.amazon.cloudformation.proxy.Logger;
 import software.amazon.cloudformation.proxy.ProgressEvent;
 import software.amazon.cloudformation.proxy.OperationStatus;
 import software.amazon.cloudformation.proxy.ResourceHandlerRequest;
+import software.amazon.cloudformation.resource.IdentifierUtils;
 
-public class ReadHandler extends BaseHandler<CallbackContext> {
+public class CreateHandler extends BaseHandler<CallbackContext> {
 
     @Override
     public ProgressEvent<ResourceModel, CallbackContext> handleRequest(
@@ -16,6 +17,14 @@ public class ReadHandler extends BaseHandler<CallbackContext> {
         final Logger logger) {
 
         final ResourceModel model = request.getDesiredResourceState();
+
+        // Set model primary ID if absent
+        if(model.getId() == null) {
+            model.setId(IdentifierUtils.generateResourceIdentifier(
+                request.getLogicalResourceIdentifier(),
+                request.getClientRequestToken()
+            ));
+        }
 
         // TODO : put your code here
