+++ aws-ssm-parameter/src/test/java/com/amazonaws/ssm/parameter/DeleteHandlerTest.java
@@ -1,4 +1,4 @@
-package software.amazon.ssm.parameter;
+package com.amazonaws.ssm.parameter;
 
 import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
 import software.amazon.cloudformation.proxy.Logger;
