+++ aws-ssm-parameter/src/test/java/com/amazonaws/ssm/parameter/UpdateHandlerTest.java
@@ -1,5 +1,6 @@
-package software.amazon.ssm.parameter;
+package com.amazonaws.ssm.parameter;
 
+import org.assertj.core.api.Assertions;
 import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
 import software.amazon.cloudformation.proxy.Logger;
 import software.amazon.cloudformation.proxy.OperationStatus;
@@ -47,7 +48,7 @@ public class UpdateHandlerTest {
         assertThat(response.getCallbackContext()).isNull();
         assertThat(response.getCallbackDelaySeconds()).isEqualTo(0);
         assertThat(response.getResourceModel()).isEqualTo(request.getDesiredResourceState());
-        assertThat(response.getResourceModels()).isNull();
+        Assertions.assertThat(response.getResourceModels()).isNull();
         assertThat(response.getMessage()).isNull();
         assertThat(response.getErrorCode()).isNull();
     }
