+++ aws-ssm-parameter/template.yml
@@ -10,14 +10,13 @@ Resources:
   TypeFunction:
     Type: AWS::Serverless::Function
     Properties:
-      Handler: software.amazon.ssm.parameter.HandlerWrapper::handleRequest
+      Handler: com.amazonaws.ssm.parameter.HandlerWrapper::handleRequest
       Runtime: java8
       CodeUri: ./target/aws-ssm-parameter-handler-1.0-SNAPSHOT.jar
 
   TestEntrypoint:
     Type: AWS::Serverless::Function
     Properties:
-      Handler: software.amazon.ssm.parameter.HandlerWrapper::testEntrypoint
+      Handler: com.amazonaws.ssm.parameter.HandlerWrapper::testEntrypoint
       Runtime: java8
       CodeUri: ./target/aws-ssm-parameter-handler-1.0-SNAPSHOT.jar
-
